requirements.txt
Django~=3.1.8
djangorestframework~=3.11.2
gunicorn